fix module exports.
